magnet added, yare yare

+++ src/coin.cpp
@@ -1,14 +1,14 @@
 #include "coin.h"
 #include "main.h"
 #define GLM_ENABLE_EXPERIMENTAL
-Coin::Coin(float radius, float x, float y, color_t color) {
+Coin::Coin(double radius, double x, double y, color_t color) {
     // Our vertices. Three consecutive floats give a 3D vertex; Three consecutive vertices give a triangle.
     // A cube has 6 faces with 2 triangles each, so this makes 6*2=12 triangles, and 12*3 vertices
     this->position = glm::vec3(x, y, 0);
     this -> rotation = 0;
     this -> isdraw = true;
     GLfloat vertex_buffer_data[9 * 300];
-    float angle = (2 * PI) / 300, theta = 0;
+    double angle = (2 * PI) / 300, theta = 0;
     for(int i = 0; i < 9 * 300; i++)
     {
         if(i % 3 == 2 || i % 9 < 3)
