+++ src/coin.h
@@ -6,7 +6,7 @@
 class Coin {
 public:
     Coin() {}
-    Coin(float radius, float x, float y, color_t color);
+    Coin(double radius, double x, double y, color_t color);
     float rotation;
     glm::vec3 position;
     void draw(glm::mat4 VP);
